S3AFileRegionFormat support. (#4)
As part of the Provided Storage PoC, we can store the block.csv in S3a. This
means the blocks.csv file is available on S3. This is a fast way to get Provided
Storage working on a cluster that doesn't update the fsimage but it's wholly
inadequate for the ultimate release since S3 isn't a key value store. If you
would like a key value store in production system using S3, consider using
ZookeeperFileRegionFormat.